refactor defaultValues endpoint
